# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ERP Van Sales API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchases:
    get:
      operationId: listPurchases
      tags: [purchases]
      summary: List purchase orders
      responses:
        "200":
          description: List of purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Create purchase order (Bon d'achat)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseBody"
      responses:
        "201":
          description: Purchase created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}:
    get:
      operationId: getPurchase
      tags: [purchases]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Get purchase
      responses:
        "200":
          description: Purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}/payment:
    post:
      operationId: addPurchasePayment
      tags: [purchases]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Record payment for purchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchasePaymentBody"
      responses:
        "200":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"

  # CLIENTS
...
  /returns:
    get:
      operationId: listReturns
      tags: [returns]
      summary: List returns
      responses:
        "200":
          description: List of returns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Return"
    post:
      operationId: createReturn
      tags: [returns]
      summary: Create return
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReturnBody"
      responses:
        "201":
          description: Return created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Return"
  /returns/{id}:
    get:
      operationId: getReturn
      tags: [returns]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Get return
      responses:
        "200":
          description: Return
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Return"
...
  /reports/daily:
    get:
      operationId: getDailyReport
      tags: [reports]
      summary: Get daily sales report
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: truckId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
  /reports/trucks:
    get:
      operationId: getTruckPerformance
      tags: [reports]
      summary: Get truck performance report
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Truck performance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TruckPerformance"
...
  /reports/top-products:
    get:
      operationId: getTopProducts
      tags: [reports]
      summary: Get top selling products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
components:
  schemas:
...
    Purchase:
      type: object
      properties:
        id:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: string
        totalAmount:
          type: number
        paidAmount:
          type: number
        remainingAmount:
          type: number
        paymentStatus:
          type: string
          enum: [pending, partial, paid]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItem"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - supplierId
        - supplierName
        - totalAmount
        - paidAmount
        - remainingAmount
        - paymentStatus
        - items
        - createdAt
    PurchaseItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        purchasePrice:
          type: number
        subtotal:
          type: number
      required:
        - id
        - productId
        - productName
        - quantity
        - purchasePrice
        - subtotal
    CreatePurchaseBody:
      type: object
      properties:
        supplierId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItemInput"
        initialPayment:
          type: number
      required:
        - supplierId
        - items
    PurchaseItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: number
        purchasePrice:
          type: number
      required:
        - productId
        - quantity
        - purchasePrice
    PurchasePaymentBody:
      type: object
      properties:
        amount:
          type: number
      required:
        - amount
...
    Return:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [client_return, truck_return]
        truckId:
          type: integer
          nullable: true
        truckName:
          type: string
          nullable: true
        clientId:
          type: integer
          nullable: true
        clientName:
          type: string
          nullable: true
        invoiceId:
          type: integer
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReturnItem"
        totalAmount:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - items
        - totalAmount
        - createdAt
    ReturnItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        subtotal:
          type: number
      required:
        - productId
        - productName
        - quantity
        - unitPrice
        - subtotal
    CreateReturnBody:
      type: object
      properties:
        type:
          type: string
          enum: [client_return, truck_return]
        truckId:
          type: integer
          nullable: true
        clientId:
          type: integer
          nullable: true
        invoiceId:
          type: integer
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReturnItemInput"
      required:
        - type
        - items
    ReturnItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: number
        unitPrice:
          type: number
      required:
        - productId
        - quantity
        - unitPrice
...
    DailyReport:
      type: object
      properties:
        date:
          type: string
        totalSales:
          type: number
        cashSales:
          type: number
        creditSales:
          type: number
        totalCommission:
          type: number
        invoiceCount:
          type: integer
        topProducts:
          type: array
          items:
            $ref: "#/components/schemas/TopProduct"
        byTruck:
          type: array
          items:
            $ref: "#/components/schemas/TruckPerformance"
      required:
        - date
        - totalSales
        - cashSales
        - creditSales
        - totalCommission
        - invoiceCount
        - topProducts
        - byTruck
    TruckPerformance:
      type: object
      properties:
        truckId:
          type: integer
        truckName:
          type: string
        vendeurName:
          type: string
          nullable: true
        totalSales:
          type: number
        cashSales:
          type: number
        creditSales:
          type: number
        totalCommission:
          type: number
        invoiceCount:
          type: integer
      required:
        - truckId
        - truckName
        - totalSales
        - cashSales
        - creditSales
        - totalCommission
        - invoiceCount
...
    TopProduct:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        totalQuantity:
          type: number
        totalRevenue:
          type: number
        invoiceCount:
          type: integer
      required:
        - productId
        - productName
        - totalQuantity
        - totalRevenue
        - invoiceCount
...